fixed a bug with state. now each input has its own state. removed GoalList component as well.
